# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/foo": {
            "get": {
                "description": "List foos filtered by org and namespace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Foo"
                ],
                "summary": "List foos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FooListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new foo resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Foo"
                ],
                "summary": "Create foo",
                "parameters": [
                    {
                        "description": "Foo payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateFooDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.FooDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    }
                }
            }
        },
        "/api/v1/foo/{id}": {
            "get": {
                "description": "Fetch a foo by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Foo"
                ],
                "summary": "Get foo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FooDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a foo by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Foo"
                ],
                "summary": "Update foo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Foo update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateFooDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FooDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a foo by ID",
                "tags": [
                    "Foo"
                ],
                "summary": "Delete foo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Problem"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.CreateFooDTO": {
            "type": "object",
            "required": [
                "name",
                "namespace",
                "org_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "org_id": {
                    "type": "string"
                }
            }
        },
        "types.FooDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "01HZJ8K9M2N3P4Q5R6S7T8U9V"
                },
                "name": {
                    "type": "string",
                    "example": "my-foo"
                },
                "namespace": {
                    "type": "string",
                    "example": "default"
                },
                "org_id": {
                    "type": "string",
                    "example": "org-123"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "types.FooListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.FooDTO"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/types.ListMeta"
                }
            }
        },
        "types.ListMeta": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "name is required"
                },
                "instance": {
                    "type": "string",
                    "example": "/api/v1/foo/123"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "title": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "type": {
                    "type": "string",
                    "example": "https://example.com/validation-error"
                }
            }
        },
        "types.UpdateFooDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}